Add single character Parser correction
If a scan of an account number is found to be illegible or erroneous and
can be changed to a valid account number by adding or removing just one
character, then make that change and call it a day.

If there are multiple valid corrections of a different character each,
then fail the parse as "ambiguous".

Otherwise, parse does what it did before.